fix: move constant and flags to prevent loop
