+++ pkg/consts/consts.go
@@ -54,4 +54,6 @@ const (
 	ImageAnnotationKey      = "hauler.dev/key"
 	ImageAnnotationPlatform = "hauler.dev/platform"
 	ImageAnnotationRegistry = "hauler.dev/registry"
+
+	DefaultStoreName = "store"
 )
